Reuse common hash table code
